Move task section components into subfolder
