+++ frontend/src/components/AgencyProfile/TaskSection/AgencyTaskItem.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./AgencyProfile.css";
+import "../AgencyProfile.css";
 
 function AgencyTaskItem({ title, dueDate, index, color, onClick }) {
   return (
