+++ frontend/src/components/AgencyProfile/TaskSection/AgencyTaskSection.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import AgencyTaskColumn from "./AgencyTaskColumn";
-import "./AgencyProfile.css";
+import "../AgencyProfile.css";
 
 function AgencyTaskSection({ taskList, onEditTask, onCreateTask }) {
   let pendingTasks = [];
